raftserver: move some codes to sub server mod.
